LIBFCREPO-1333. Added "has_updated" method to Vocabulary model
Added "has_updated" method to Vocabulary model that returns True when
a Vocabulary (or one of its dependents) has unpublished changes.

Testing this method revealed an interaction between then "modified" and
"published" time, where the act of saving the "published" timestamp
sets the "modified" timestamp to a time a few milliseconds later,
throwing off the "has_updated" check. Fixed this by setting both the
"published" and "modified" timestamps to the same timestamp (via a
direct database update) when the vocabulary is published.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1333