LIBFCREPO-1334. Enable HTMX "useTemplateFragments"
The "hx-swap-oob" directive has problems with responses returning
HTML <table> fragments. Enabling the HTMX "useTemplateFragments"
attribute appears to fix the issue (originally suggested in
https://stackoverflow.com/a/77255665).

Note: The "useTemplateFragments" directive does not work with
Internet Explorer 11, but it was retired in June, 2022, replaced with
Microsoft Edge, so presumably we don't need to worry about it.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1333