LIBFCREPO-1334. Add "Publish Updates" button
Implemented using a “PublishUpdatesMixin” class, that can be added to
the various Django views that trigger changes to the Vocabulary, Term,
or Property models:

* TermsView
* TermView
* PropertyView
* PropertyEditView

Because the “PublishUpdatesMixin” can be called from a variety of views,
the logic for determining whether the vocabulary has publishable updates
has been encapsulated in the “vocabulary_has_updated” method, which does
the following:

* Any DELETE or POST request is assumed to be an update. This is needed
  because we can't retrieve the object on DELETE requests, and some
  POSTS (such as from TermsView) don't use templates, and so don't
  respond to "self.get_object".
* All other requests are assume to respond to “self.get_object”. If the
  returned object is a Property, Term, or Vocabulary instance, the
  “has_updated” method on the associated vocabulary object is called
* All other cases return False

The PublishUpdatesMixin renders the
“src/vocabs/templates/vocabs/publish_updates_button.html” template, to
show the button (passing a “show_publish_updates_button” variable in
template context), appending it on to whatever content is already being
returned by the view.

The template uses the HTMX “hx-swap-oob” directive to piggy-back the
display of the “Publish Updates” button onto any content change, so that
the “Publish Updates” button is displayed in the correct location.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1334