LIBFCREPO-1359. Prevent HX-Trigger events from being overwritten
Fix an issue where the "grove:vocabUpdated" event was overwriting the
"grove:termAdded" event in the "HX-Trigger" response header. This was
leading to the "Term" form not being replaced with an empty form when
a new term is added.

This fix, which may not be optimal, or scale for additional events,
essentially sequences the "grove:termAdded" and "grove:vocabUpdated"
events, so only one is processed at a time. When a new Term is added
this results in one POST (the term submission), and two GET events (the
new Term form request, and the vocab updated request.

It is possible to add multiple events to the HX-Trigger handler (see the
"Multiple Triggers" section in https://htmx.org/headers/hx-trigger/),
but in that case there are three GET requests (beyond the initial POST):

* A GET for the "grove:vocabUpdated" event
* A GET for the "grove:termAdded" event, with a second GET request from
  the "PublishUpdatesMixin" dispatch method because the vocabulary is
  still updated.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1359